requirements.txt
pyupbit>=0.2.32
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
ta>=0.11.0
requests>=2.31.0
matplotlib>=3.7.0
